Included MinGW 4.6.*
This is a substitute for those that prefer to use c++ or g++ via ada.
MinGW allows us to compile the project via CLI, adjacent to working on
the ada dedicated server.  However, do keep in mind that with this
compiler - we are bound to Windows platform.  With that, one can not
compile the binary in the Windows OS and try to execute the build in the
*nix environment, adjacent for the other-way around.  Furthermore, we
might be able to create some shell scripts to levy the make file, but
we'll come across this later.  For those using Code::Blocks, Visual
Studio, or any other IDE, you are free to use any IDE's that you like.