Added comment for BK
Please update your code to append to our model.  The code the professor
provided does not append to our model as our structure is not directly
from the instructor's examples.  Besides to clinging to three mass
pointers, we're using primarily only one.  The search function will have
a total of three nodes but two scoped locally for proper indexing.