Expunged deprecated code && revised update function
Thrashed the older and deprecated function
'Update_Passenger_Information()', this has been replenished by the newly
added function 'UpdatePassengerInformation()'.  This newly function is
slimer and should cover everything right out the box.  Though this is
NOT a healthy and optimized way of doing things, but it should do
essentially what we need regardless.