Changed program to OOP
I added all of the inline functions to a class called Reservation.  I
used head as a class variable and got rid of all the head parameters.  I
moved all of the functions to the linkedlist.cpp file, and all the
declarations to the header file.  I tested the code and it seemed to
work the same after I made the changes.  I also updated the main to
reflect the oop changes.