Updated Delete Function and fixed issue with print list function
I update the delete function to work with the search function and
implemented it in the menu.  The delete function should be fully
functional now (unless I missed something I guess).  I also fixed an
issue I caused in the print passenger list function where the function
would actually update the head pointer as it printed.