Fixed LNK2005 error [VS only] && more minor cleanups
It seems the inclusion of the impelementation is causing the linker to
do some crazy procedures, with that - I have commented it out as VS will
do this for me due to the OOP merger.  If anyone else _/DEPENDS/_ on
that one call, please let me know and I will try to work with you on a
revert.