Added Print Check In List Function and fixed issues with check in function
I added the print check in list function and fixed some input issues
with the check in functionality.  I should also mention that in this
version I have added an include for  cstdlib (for the rand functions) to
the LinkedList.cpp file for error checking purposes and I had to
uncomment the linkedlist.cpp include in the driver to get it to compile
for me.