Updated notes to dismiss MinGW
Revised the notes to include the compilers offered on ADA, c++ && g++ via extCMD's in Linux's shell.  We will not be using MinGW as "Something Happened.".  I am still unsure what caused the binary to disappear or maybe I did not install it properly when I installed the 4.6.x build?  Either way, we will be using the extCMD's on ada.