Expunged the Tools directory and all sub-contents.
Thrashed all of the contents within the Tools directory and also
expunged the directory itself.  This directory was initially for MinGW,
but as we are unable to use it presently, no point keeping it around.
Additionally, we can compact the data ourselfs without needing 7Zip and
a batch script to do it.  Thus, all of the contents within that
directory were merely superfluous to us.