Added output
Added all of the output as of revision
62f131b191dad7e8fcf6956d03f7a27f5c702900 .
With this, all (or mostly) functionalities possible within this program
has been tested and provided within the 'output.txt' file.
If incase needed, I have also provided Output_key.txt, which hosts what
was inputted into the program during the test run.